--- src/DocumentActionButtons.scss
@@ -1,5 +1,0 @@
-@import "_variables";
-button.btn.btn-success.paid-button:hover{
-    background-color: transparent;
-    color: $success;
-}
