+++ src/DocumentActions/DocumentActionButtons.scss
@@ -1,0 +1,5 @@
+@import "../_variables";
+button.btn.btn-success.paid-button:hover {
+  background-color: transparent;
+  color: $success;
+}
