+++ src/DocumentActions/index.js
@@ -1,0 +1,5 @@
+export {
+  DocumentActionsContext,
+  DocumentActionsProvider,
+} from "./DocumentActionsContext";
+export { DocumentActionButtons } from "./DocumentActionButtons";
