+++ src/DocumentViewer/index.js
@@ -1,1 +1,2 @@
+export { DocumentViewerTab } from "./DocumentViewerTab";
 export { DocumentViewer } from "./DocumentViewer";
