+++ src/index.scss
@@ -61,7 +61,7 @@ input[type="number"] {
   }
 }
 
-// Common page settings for DocumentViewer
+// Common page settings for DocumentViewerTab
 
 .page {
   position: relative;
